group-5/project-1/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
requests>=2.31
pdfplumber>=0.10
python-docx>=1.0
dashscope>=1.14
